Split Take and Take(Range) files
